query-time road class filtering
