+++ rust/routee-compass-core/src/model/frontier/default/mod.rs
@@ -1,2 +1,1 @@
 pub mod no_restriction;
-pub mod road_class;
