+++ rust/routee-compass/src/app/compass/config/frontier_model/mod.rs
@@ -1,2 +1,2 @@
 pub mod no_restriction_builder;
-pub mod road_class_builder;
+pub mod road_class;
