+++ rust/routee-compass/src/app/compass/config/frontier_model/road_class/mod.rs
@@ -1,0 +1,3 @@
+pub mod road_class_builder;
+pub mod road_class_model;
+pub mod road_class_service;
